tensorflow-rock/requirements.txt
tensorflow-gpu
matplotlib
numpy==1.16
pandas
scipy